refactor: add assert for id

+++ src/app/presentation/controllers/photoController.js
@@ -38,6 +38,8 @@ router.delete('/:id', (req, res) => {
 	let photoId = req.params.id;
 	console.log(` 🛃 [PhotoController] delete() #id: `, photoId);
 
+	if (!assertIsValidString(photoId)) { return sendAssertion(res, 'id'); }
+
 	unsplashService.deletePhoto(photoId);
 });
